refactor: user repository 제외
