+++ src/core/auth/jwt.strategy.ts
@@ -1,13 +1,21 @@
-import { Injectable, UnauthorizedException } from '@nestjs/common';
+import {
+    forwardRef,
+    Inject,
+    Injectable,
+    UnauthorizedException,
+} from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { Request } from 'express';
-import { UserRepository } from '../users/repository/user.repository';
 import { UserEntity } from '../users/entities/user.entity';
+import { UserService } from '../users/user.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-    constructor(private readonly userRepository: UserRepository) {
+    constructor(
+        @Inject(forwardRef(() => UserService))
+        private readonly userService: UserService,
+    ) {
         super({
             jwtFromRequest: ExtractJwt.fromExtractors([
                 (request: Request): string | null => {
@@ -20,7 +28,7 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
 
     async validate(payload: any): Promise<UserEntity> {
         // 토큰에서 사용자 ID 추출 후 사용자 정보 검증
-        const user = await this.userRepository.findById(payload.id);
+        const user = await this.userService.findById(payload.id);
         if (!user) {
             throw new UnauthorizedException('존재하지 않는 사용자입니다.');
         }
