+++ src/core/users/user.service.module.ts
@@ -5,7 +5,6 @@ import { ResumeServiceModule } from '../resumes/resume.service.module';
 import { HttpModule } from '@nestjs/axios';
 import { TaskServiceModule } from '../../core/task/taskService.module';
 import { UserExperienceServiceModule } from '../userExperiences/userExperience.service.module';
-import { UserRepository } from './repository/user.repository';
 
 @Global()
 @Module({
@@ -15,7 +14,7 @@ import { UserRepository } from './repository/user.repository';
         forwardRef(() => ResumeServiceModule),
         UserExperienceServiceModule,
     ],
-    providers: [UserService, UserRepository, PrismaService],
-    exports: [UserService, UserRepository],
+    providers: [UserService, PrismaService],
+    exports: [UserService],
 })
 export class UserServiceModule {}
